Work on adding arcanist JUnit coverage and testing
Summary:
In order to allow CI/CD integration with the react host, and to improve code quality, we need to integrate Arcanist as well as Jest JUnit/Cobertura/Clover test and coverage reporting.

This diff adds initial support for this.

Test Plan: Adding another test in a new diff after this diff lands should increase coverage.

Reviewers: mr-asher, TomReichardt

Reviewed By: mr-asher

Differential Revision: https://phab.gw-cloud.org/D19